Multi CPU scaling policies
This change allows CPU threshold to be set in a more flexible way.
So far the threshold was hardcoded to 100% but now it's possible to provide values in the deployment manifest.
This is very useful if diego VMs have more than one CPU core and you want your app to scale when CPU usage is reported as more than 100%

In the nodejs app, multiple workers are now spun up to utilise multiple CPU
threads when trying to achieve a load of more than 100% CPU.
It is now possible to configure the tests to check that an app is able to scale up if CPU usage exceeds
an upper CPU threshold of over 100%.

(cherry picked from commit 3e0781c3aca8dd37aacd5f7db696b17f0560c691)