Add missing linters and merge existing linters to save on runners
The requirement for the linters to pass will be enforced implicitly by
requiring converstions to be resolved, as the reviewdog linters
comment.

This PR merges previously separate workflows, as the linters are not
configured to error out in case violations are found, which means we can
cut down on runner usage.