Reducing the `prConcurrentLimit` to `2`
As runners are scarce in the `cloudfoundry` org,
we want to limit the number of concurrent PRs.

Every renovate PR gets rebased after every commit to main for
[good reasons](https://docs.renovatebot.com/configuration-options/#rebasewhen).

This means if there are the default `10` PRs open concurrently,
merging one PR currently triggers at least 9 * 31 = 279 workflows.

I chose the value of `2` so that if one PR fails to validate
consistently there's a second slot for other PRs. If two PRs
fail to validate consistently, probably a deeper look is necessary
anyways.