feat(scalingengine): Support label to disable autoscaling
# Background
During DevOps tasks such as deployments or periods where manual
scaling is required (e.g., in an outage situation) a simple and
quick way to disable autoscaling might be helpful that does not
involve editing or deleting the autoscaling policy.

# Details
This PR introduces a label
`app-autoscaler.cloudfoundry.org/disable-autoscaling`.
If it is set on an app the app-autoscaler will not
scale it, even if a scaling policy is active.