fix(ci): Add auto-approve step for PRs by renovate[bot
# Issue

The workflow could be triggered by both pushes and PRs.
However, the PR trigger was misconfigured to require a PR against
`renovate/**`, when a renovate PR would naturally be against `main`.
This meant, that the `push` event was the one to trigger. However, in
this case the `hmarr/auto-approve-action` would fail with

> This action must be run using a `pull_request` event or have an
explicit `pull-request-number` provided

# Solution

Instead of fixing the workflow when triggered by the `push` event, it
was fixed to correctly run when triggered by the `pull_request`.