chore(ci): Switch to `ubuntu22`
# Issue
Since [PR #2905] our GitHub Actions workflows using `ruby/setup-ruby`
started failing with

```
Error: The current runner (ubuntu-24.04-x64) was detected as self-hosted because the platform does not match a GitHub-hosted runner image (or that image is deprecated and no longer supported).

```

This is due to the mismatch of the Ubuntu version configured and
actually used.

# Fix

Configure `ImageOS: "ubuntu24"`.

[PR #2905]: https://github.com/cloudfoundry/app-autoscaler-release/pull/2905