feat(broker_test): Dynamically determine plans to update
# Issue

Previously we backed some assumptions into our plan update tests:
Their must be at least two plans, the first being updatable, and another
with a specific name that should not be updatable.

This made the tests very dependent on the configuration chosen by the
platform operator.

# Fix

We now dynamically choose the first updatable plan and update to second.
If there are no two plans or none is updatable, we skip the test

# Note

The test testing that you cannot update from the first plan has been
removed, as this is already guaranteed by the Cloud Controller, so the
test was just excercising Cloud Controller code.