fix(broker_test): Do not silently skip plan update tests
# Issue

The plan update acceptance tests require requires a deployment with two
service plans, at least one of which is updatable. This is the case in
our CI, but might not be the case when you productively run
app-autoscaler.

On the other hand this test might be silently skipped if we change
something to the CI setup.

# Fix

Do not silently skip it, but require to explicitly skip it with

```
--skip "updating service plans"
```

and do that in `test_default` as this is the default set of tests we
expect customers to run.