fix(acceptance): fix naming of plan update tests names (#2938)
* feat(broker_test): Dynamically determine plans to update

# Issue

Previously we backed some assumptions into our plan update tests:
Their must be at least two plans, the first being updatable, and another
with a specific name that should not be updatable.

This made the tests very dependent on the configuration chosen by the
platform operator.

# Fix

We now dynamically choose the first updatable plan and update to second.
If there are no two plans or none is updatable, we skip the test

# Note

The test testing that you cannot update from the first plan has been
removed, as this is already guaranteed by the Cloud Controller, so the
test was just excercising Cloud Controller code.

* fix(broker_test): Do not silently skip plan update tests

# Issue

The plan update acceptance tests require requires a deployment with two
service plans, at least one of which is updatable. This is the case in
our CI, but might not be the case when you productively run
app-autoscaler.

On the other hand this test might be silently skipped if we change
something to the CI setup.

# Fix

Do not silently skip it, but require to explicitly skip it with

```
--skip "updating service plans"
```

and do that in `test_default` as this is the default set of tests we
expect customers to run.