chore(asdf2devbox): Switch to Java 21
# Issue

As the Nix package name contains the Java major version, the translation
needs to be updated now that we have switched to version 21

# Fix

Just a static string change. Parsing it from the version number would be
possible, but as it is a LTS release the next update should be in the
far future.