chore(asdf2devbox): Only update to latest if no previous version is installed
# Issue

If a version of a program was not packaged in NixOS, `asdf2devbox` would
update to latest which might increase the version skew.

E.g. Go 1.22.4 is not packaged in NixOS, updating to it would update to
`go@latest` which might be Go 1.23 - in effect a major version upgrade.

# Fix

asdf2devbox does not update if a version was already referenced in
devbox, to minimize version skew.