chore(asdf): Remove ASDF
# Issue

We previously used ASDF to manage our dev environment but meanwhile switched to devbox.

As the `devbox.json` was previously not kept up-to-date by renovate we kept ASDF's `.tool-version` files around for some handbuilt automation.

As devbox is now supported by renovate we can remove the left-overs.

# Fix

Remove all `.tool-versions` and automation built on it.