# Issue
As an app-autoscaler developer
I want to have as few dependencies as possible
So that I don't have to watch out for so many vulnerabilities in unmaintained dependencies.

# Fix

As we already have two HTTP libraries (gorilla/mux and net/http) get rid of a third one for the test app (gin).

# AI Disclaimer

Prompted by the committer, most of the work was done by GitHub copilot.